Split Repositories into separate files
Preparing for more repository types
