+++ src/main/scala/io/viash/functionality/dependencies/Repository.scala
@@ -17,12 +17,10 @@
 
 package io.viash.functionality.dependencies
 
-import io.viash.helpers.Exec
 import io.viash.helpers.IO
 import java.nio.file.Paths
 import io.viash.schemas._
 import java.nio.file.Path
-import java.io.File
 import java.nio.file.Files
 
 @description("Specifies a repository where dependency components can be found.")
@@ -126,143 +124,4 @@ object Repository {
     storeRepositoryInCache(newRepo)
     newRepo
   }
-}
-
-@description("A GitHub repository where remote dependency components can be found.")
-@example(
-  """type: github
-    |uri: openpipelines-bio/modules
-    |tag: 0.3.0
-    |""".stripMargin,
-  "yaml"
-)
-@example(
-  """name: viash-testns
-    |type: github
-    |uri: viash-io/viash
-    |tag: 0.7.1
-    |path: src/test/resources/testns
-    |""".stripMargin,
-  "yaml"
-  )
-case class GithubRepository(
-  name: String,
-
-  @description("Defines the repository as a GitHub repository.")
-  `type`: String = "github",
-
-  @description("The GitHub `organization/repository_name` of the repository.")
-  @example("uri: viash-io/viash", "yaml")
-  uri: String,
-  tag: Option[String],
-  path: Option[String] = None,
-  localPath: String = ""
-) extends Repository {
-  
-  def copyRepo(
-    name: String,
-   `type`: String,
-    tag: Option[String],
-    path: Option[String],
-    localPath: String
-  ): Repository = {
-    copy(name, `type`, uri, tag, path, localPath)
-  }
-
-  lazy val fullUri = s"git@github.com:$uri.git"
-
-  // Clone of single branch with depth 1 but without checking out files
-  def checkoutSparse(): GithubRepository = {
-    val temporaryFolder = IO.makeTemp("viash_hub_repo")
-    val cwd = Some(temporaryFolder.toFile)
-
-    Console.err.println(s"temporaryFolder: $temporaryFolder uri: $uri fullUri: $fullUri")
-
-    val singleBranch = tag match {
-      case None => List("--single-branch")
-      case Some(value) => List("--single-branch", "--branch", value)
-    }
-
-    val loggers = Seq[String => Unit] { (str: String) => {Console.err.println(str)} }
-
-    val out = Exec.runCatch(
-      List("git", "clone", fullUri, "--no-checkout", "--depth", "1") ++ singleBranch :+ ".",
-      cwd = cwd,
-      loggers = loggers,
-    )
-
-    copy(localPath = temporaryFolder.toString)
-  }
-
-  def hasBranch(name: String, cwd: Option[File]): Boolean = {
-    val out = Exec.runCatch(
-      List("git", "show-ref", "--verify", "--quiet", s"refs/heads/$name"),
-      cwd = cwd
-    )
-    out.exitValue == 0
-  }
-
-  def hasTag(name: String, cwd: Option[File]): Boolean = {
-    val out = Exec.runCatch(
-      List("git", "show-ref", "--verify", "--quiet", s"refs/tags/$name"),
-      cwd = cwd
-    )
-    out.exitValue == 0
-  }
-
-  // Checkout of files from already cloned repository. Limit file checkout to the path that was specified
-  def checkout(): GithubRepository = {
-    val pathStr = path.getOrElse(".")
-    val cwd = Some(Paths.get(localPath).toFile)
-    val checkoutName = tag match {
-      case Some(name) if hasBranch(name, cwd) => s"origin/$name"
-      case Some(name) if hasTag(name, cwd) => s"tags/$name"
-      case _ => "origin/HEAD"
-    }
-
-    val out = Exec.runCatch(
-      List("git", "checkout", checkoutName, "--", pathStr),
-      cwd = cwd
-    )
-
-    Console.err.println(s"checkout out: ${out.command} ${out.exitValue} ${out.output}")
-
-    if (path.isDefined)
-      copy(localPath = Paths.get(localPath, path.get).toString)
-    else
-      // no changes to be made
-      this
-  }
-
-  // Get the repository part of where dependencies should be located in the target/dependencies folder
-  def subOutputPath: String = Paths.get(`type`, uri, tag.getOrElse("")).toString()
-}
-
-@description(
-  """Defines a locally present and available repository.
-    |This can be used to define components from the same code base as the current component.
-    |Alternatively, this can be used to refer to a code repository present on the local hard-drive instead of fetchable remotely, for example during development.
-    |""".stripMargin
-)
-case class LocalRepository(
-  name: String = "",
-
-  @description("Defines the repository as a locally present and available repository.")
-  `type`: String = "local",
-  tag: Option[String] = None,
-  path: Option[String] = None,
-  localPath: String = ""
-) extends Repository {
-
-  def copyRepo(
-    name: String,
-    `type`: String,
-    tag: Option[String],
-    path: Option[String],
-    localPath: String
-  ): Repository = {
-    copy(name, `type`, tag, path, localPath)
-  }
-
-  def subOutputPath: String = Paths.get(`type`, tag.getOrElse("")).toString()
 }
